// src/interfaces/team-stats.ts
export interface TeamStats {
  draws: number;
  goalsAgainst: number;
  goalsFor: number;
  loses: number;
  points: number;
  teamId: string;
  wins: number;
}

export interface TableStats {
  draws: number;
  goalsAgainst: number;
  goalsFor: number;
  loses: number;
  points: number;
  wins: number;
}